change user model responses

+++ server/controllers/user_controller.py
@@ -1,5 +1,5 @@
-from flask_restplus import Resource, fields, reqparse
-from flask import request, make_response
+from flask_restplus import Resource, fields, reqparse, marshal
+from flask import request, make_response, jsonify
 from werkzeug.security import generate_password_hash, check_password_hash
 
 from server.api.restplus import api
@@ -9,6 +9,7 @@ from server.models import db
 ns = api.namespace('users', description='Operations related to users')
 
 user_dto = api.model('user', {
+    'user_uuid': fields.String(required=True, description='user uuid'),
     'email': fields.String(required=True, description='user email address'),
     'username': fields.String(required=True, description='user username'),
 })
@@ -24,7 +25,7 @@ class UserList(Resource):
             results = User.query.all()
             return results
         except Exception as e:
-            return make_response({"message": str(e)}, 500)
+            return {"message": str(e)}, 500
 
 
 @ns.route('/add')
@@ -51,9 +52,9 @@ class AddUser(Resource):
             db.session.add(newUser)
             db.session.commit()
         except Exception as e:
-            return make_response({"message": str(e)}, 500)
+            return {"message": str(e)}, 500
 
-        return make_response({'message': 'user has been created successfully.'}, 201)
+        return {'message': 'user has been created successfully.'}, 201
 
 
 @ns.route('/delete')
@@ -73,27 +74,34 @@ class DeleteUser(Resource):
 
         try:
             userToBeDeleted = User.query.filter_by(username=username).first()
-            db.session.delete(userToBeDeleted)
-            db.session.commit()
+            if userToBeDeleted:
+                db.session.delete(userToBeDeleted)
+                db.session.commit()
+            else:
+                return {'message': 'user not found.'}, 404
         except Exception as e:
-            return make_response({"message": str(e)}, 500)
+            return {"message": str(e)}, 500
 
-        return make_response({'message': 'user has been deleted successfully.'}, 201)
+        return {'message': 'user has been deleted successfully.'}, 201
 
 
 @ns.route('/<string:username>')
 class UserSearch(Resource):
-    @ns.marshal_list_with(user_dto)
+    @ns.response(code=201, model=user_dto, description='Success')
+    @ns.response(code=404, description='Not Found')
     def get(self, username):
         """
         Gets a specified user
         """
         try:
             queriedUser = User.query.filter_by(username=username).first()
-            return queriedUser
+            if queriedUser:
+                return marshal(queriedUser, user_dto)
+            else:
+                return {"message": 'user not found'}, 404
 
         except Exception as e:
-            return make_response({"message": str(e)}, 500)
+            return {"message": str(e)}, 500
 
     user_edit_parser = reqparse.RequestParser()
     user_edit_parser.add_argument('new_email', nullable=True, required=False, type=str, help='new email of user')
@@ -118,16 +126,46 @@ class UserSearch(Resource):
                 if data.get('new_password'):
                     userToBeEditted.password_hash = generate_password_hash(data.get('new_password'))
             else:
-                return make_response({'message': 'user specified not found in database'}, 201)
+                return {'message': 'user specified not found in database'}, 201
 
             db.session.commit()
         except Exception as e:
-            return make_response({"message": str(e)}, 500)
+            return {"message": str(e)}, 500
 
-        return make_response({'message': 'user has been edited successfully.'}, 201)
+        return {'message': 'user has been edited successfully.'}, 201
 
 
+@ns.route('/login')
+class UserLogin(Resource):
 
+    user_login_parser = reqparse.RequestParser()
+    user_login_parser.add_argument('username', required=True, type=str, help='username of user')
+    user_login_parser.add_argument('password', required=True, type=str, help='password of user')
+
+    @ns.expect(user_login_parser)
+    @ns.response(code=201, model=user_dto, description='Success')
+    @ns.response(code=404, description='Not Found')
+    def get(self):
+        """
+        Login a specified user
+        """
+        data = request.args
+
+        input_username = data.get('username')
+        input_password = data.get('password')
+
+        try:
+            queriedUser = User.query.filter_by(username=input_username).first()
+            if queriedUser:
+                if check_password_hash(queriedUser.password_hash, input_password):
+                    return marshal(queriedUser, user_dto)
+                else:
+                    return {'message': 'invalid password'}, 404
+            else:
+                return {'message': 'username not found'}, 404
+
+        except Exception as e:
+            return {"message": str(e)}, 404
